dev-requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
fastapi==0.109.0
filelock==3.13.1
flake8==7.0.0
greenlet==3.0.3
h11==0.14.0
identify==2.5.33
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.4
mccabe==0.7.0
mypy==1.8.0
mypy-extensions==1.0.0
nodeenv==1.8.0
orjson==3.9.12
platformdirs==4.2.0
pre-commit==3.6.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydantic==2.6.0
pydantic-settings==2.1.0
pydantic_core==2.16.1
pyflakes==3.2.0
python-dotenv==1.0.1
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
virtualenv==20.25.0
